% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_metrics.R
\name{aggregate_metrics}
\alias{aggregate_metrics}
\title{Combine all numeric data from metrics into single table}
\usage{
aggregate_metrics(division_metric_list)
}
\arguments{
\item{division_metric_list}{List by division with sub-lists for each metric}
}
\value{
Table with returns for each metric for each month
}
\description{
Function to pull together reported figures from all divisions into single table
}
