% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{as.percent}
\alias{as.percent}
\title{Convert number to percentage}
\usage{
as.percent(number, digits = 1, format = "f", ...)
}
\arguments{
\item{number}{Numeric variable}
}
\value{
Percentage as character
}
\description{
Function to convert input number to percentage format
}
