% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{as_percent}
\alias{as_percent}
\title{Convert value to percentage}
\usage{
as_percent(value, digits = 0, format = "f", ...)
}
\arguments{
\item{value}{Numeric variable}

\item{digits}{Number of decimal places}

\item{format}{format}
}
\value{
Percentage as character
}
\description{
Function to convert input value to percentage format
}
