% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{assign_division_names}
\alias{assign_division_names}
\title{Assign names to list elements}
\usage{
assign_division_names(list, new_names)
}
\arguments{
\item{list}{Any list}

\item{new_names}{Names as values}
}
\value{
List named elements
}
\description{
Function to name elements of list
}
