% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{assign_list_names}
\alias{assign_list_names}
\title{Extract (single) division name}
\usage{
assign_list_names(list, new_names)
}
\arguments{
\item{list}{Any list}

\item{new_names}{Names as character vector of same length as list}

\item{file_paths}{File path as character}

\item{config_file}{.yaml file containing config, with default settings (data directory, null values, metir ranges and reporting months)}
}
\value{
Division name as character

List named elements
}
\description{
Function to extract division name (acronym) from file path

Function to name elements of list
}
