% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_metrics.R
\name{combine_metrics}
\alias{combine_metrics}
\title{Combine all data from metrics into single table}
\usage{
combine_metrics(division_metric_list)
}
\arguments{
\item{division_metric_list}{List by division with sub-lists for each metric}
}
\value{
Tidy data frame with all data for all divisions
}
\description{
Function to pull together detailed data from all divisions into single tidy data frame
}
