% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_risks.R
\name{count_risks}
\alias{count_risks}
\title{Count number of quality risks per division}
\usage{
count_risks(merged_risks)
}
\arguments{
\item{merged_risks}{List of risk tables by division}
}
\value{
List with number of risks for each division
}
\description{
Function to count number of quality risks in each division from list of risks by division
}
