% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_risks.R
\name{create_all_div_dimension_table}
\alias{create_all_div_dimension_table}
\title{Create single division/dimension table - numbers only}
\usage{
create_all_div_dimension_table(all_div_risks_merged)
}
\arguments{
\item{all_div_risks_merged}{List of merged risks by division}
}
\value{
Table of totals by division with added total column
}
\description{
Function to create a single table with dimension breakdown by division, and sum for each dimension
}
