% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_risks.R
\name{create_all_div_risk_table}
\alias{create_all_div_risk_table}
\title{Number of Quality Risks by division in table}
\usage{
create_all_div_risk_table(all_div_risks_merged)
}
\arguments{
\item{all_div_risks_merged}{List of tables for each division with merged quality risks}
}
\value{
Table
}
\description{
Function to present number of quality risks in table from list of totals by division
}
