% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_risks.R
\name{create_dim_table}
\alias{create_dim_table}
\title{Create single division/dimension table - numbers only}
\usage{
create_dim_table(merged_risks)
}
\arguments{
\item{merged_risks}{List of risk tables by division}
}
\value{
Table of risks by division and division with added "total" column
}
\description{
Function to create a single table with dimension breakdown by division, and sum for each dimension
}
