% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_risks.R
\name{create_risk_table}
\alias{create_risk_table}
\title{Number of Quality Risks by division in table}
\usage{
create_risk_table(dim_table)
}
\arguments{
\item{merged_risks}{List of risk tables by division}
}
\value{
Table of totals by division
}
\description{
Function to extract number of quality risk from list of totals by division
}
