% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_metrics.R
\name{extract_all_metrics_unnamed}
\alias{extract_all_metrics_unnamed}
\title{Extract all metric tables from files}
\usage{
extract_all_metrics_unnamed(
  file_paths,
  config_file = "D:/Repos/quality_improvement_plans/config.yaml"
)
}
\arguments{
\item{file_paths}{Character vector of file paths}

\item{config_file}{.yaml file containing config, with default settings (data directory, null values, metric ranges etc)}
}
\value{
List with sub-lists
}
\description{
Function to extract all metric tables from all excel files, returning a list of metrics tables in a list by division
}
