% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_metrics.R
\name{extract_metrics_from_tabs}
\alias{extract_metrics_from_tabs}
\title{Extract all metric tables from file}
\usage{
extract_metrics_from_tabs(
  file_path,
  config_file = "D:/Repos/quality_improvement_plans/config.yaml"
)
}
\arguments{
\item{file_path}{File path for single division}

\item{config_file}{.yaml file containing config, with default settings (data directory, null values, metric ranges etc)}
}
\value{
List of 4 tables, one for each metric
}
\description{
Function to extract metric tables from divisional excel file
}
