% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_risks.R
\name{extract_n_risk_division}
\alias{extract_n_risk_division}
\title{Count number of quality risks per division}
\usage{
extract_n_risk_division(all_div_risks_merged)
}
\arguments{
\item{all_div_risks_merged}{List of tables for each division with merged quality risks}
}
\value{
List with number of risks for each division
}
\description{
Function to count number of quality risks in each division from list of risks by division
}
