% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_risks.R
\name{extract_risks_all_div}
\alias{extract_risks_all_div}
\title{Extract Quality Risks for all divisions}
\usage{
extract_risks_all_div(file_path)
}
\arguments{
\item{file_path}{List of file paths}
}
\value{
List of tables (list by division, sub-lists by risk)
}
\description{
Function to extract tab names (for each division) and extract Quality Risk tables for all divisions
}
