% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_metrics.R
\name{get_all_merged_metric_tables}
\alias{get_all_merged_metric_tables}
\title{Merge metrics for each division into one table}
\usage{
get_all_merged_metric_tables(
  all_metric_list,
  config_file = "D:/Repos/quality_improvement_plans/config.yaml"
)
}
\arguments{
\item{all_metric_list}{List by division containing 4 metric tables for each division}

\item{config_file}{.yaml file containing config, with default settings (data directory, null values, metric ranges etc)}
}
\value{
List containing one table for each division
}
\description{
Function to merge all 4 metrics into one table for all divisions in list
}
