% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_metrics.R
\name{get_all_metrics_list}
\alias{get_all_metrics_list}
\title{Extract all metric tables from files (with named list!)}
\usage{
get_all_metrics_list(
  file_paths,
  new_list_names,
  config_file = "D:/Repos/quality_improvement_plans/config.yaml"
)
}
\arguments{
\item{file_paths}{Character vector of file paths}

\item{new_list_names}{New names as values}

\item{config_file}{.yaml file containing config, with default settings (data directory, null values, metric ranges etc)}
}
\value{
Named list with sub-lists
}
\description{
Function to extract all metric tables from all excel files, returning a list of metrics tables in a named list by division
}
