% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_risks.R
\name{get_all_risks_list}
\alias{get_all_risks_list}
\title{Extract Quality Risks for all divisions (with named list!)}
\usage{
get_all_risks_list(
  file_paths,
  new_list_names,
  config_file = "D:/Repos/quality_improvement_plans/config.yaml"
)
}
\arguments{
\item{file_paths}{Character vector of file paths}

\item{new_list_names}{New names as values}

\item{config_file}{.yaml file containing config, with default settings (data directory, null values, metric ranges etc)}
}
\value{
List of tables (list by division, sub-lists of tibbles for each risk)
}
\description{
Function to extract tab names (for each division), extract Quality Risk tables for all divisions and name list elements
}
