% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_division_name}
\alias{get_division_name}
\title{Extract single division name}
\usage{
get_division_name(file_path)
}
\arguments{
\item{file_path_list}{File paths}
}
\value{
Division name as values
}
\description{
Function to extract division name (acronym) from file path
}
