% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_division_name_list}
\alias{get_division_name_list}
\title{Extract division names}
\usage{
get_division_name_list(
  config_file = "D:/Repos/quality_improvement_plans/config.yaml",
  file_path_list
)
}
\arguments{
\item{file_path_list}{List of file paths}
}
\value{
Division names as values
}
\description{
Function to extract division names (acronyms) from file paths
}
