% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_division_names}
\alias{get_division_names}
\title{Extract division names}
\usage{
get_division_names(
  file_paths,
  config_file = "D:/Repos/quality_improvement_plans/config.yaml"
)
}
\arguments{
\item{file_paths}{List of file paths as character vector}

\item{config_file}{.yaml file containing config, with default settings (data directory, null values, metric ranges etc)}
}
\value{
All division names as character vector
}
\description{
Function to extract division names (acronyms) from file paths
}
