% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_file_path_list}
\alias{get_file_path_list}
\title{Extract file list}
\usage{
get_file_path_list(
  config_file = "D:/Repos/quality_improvement_plans/config.yaml"
)
}
\arguments{
\item{config_file}{Yaml file containing config}
}
\value{
List of file paths
}
\description{
Function to extract list of file paths from folder with QIPs, sorted alphabetically
}
