% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_risks.R
\name{get_final_risk_dimension_table}
\alias{get_final_risk_dimension_table}
\title{Table of Quality risks by division and dimension}
\usage{
get_final_risk_dimension_table(all_div_risks_merg)
}
\arguments{
\item{all_div_risks_merg}{List of tables for each division with merged quality risks}
}
\value{
Table
}
\description{
Function to produce table with all risks by dimension and division, with final % of Quality risks affecting each division
}
