% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_risks.R
\name{get_merged_risks_all_div}
\alias{get_merged_risks_all_div}
\title{Collate risks in tables for each division (list elements names)}
\usage{
get_merged_risks_all_div(all_div_sep_risk_list_named, division_names)
}
\arguments{
\item{all_div_sep_risk_list_named}{List by division, sub-list by risk}

\item{division_names}{Names as values}
}
\value{
List of tables (named)
}
\description{
Function to produce table of risks for each division, with totals for risk to each dimension and named list elements
}
