% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_risks.R
\name{get_risk_dimension_tables}
\alias{get_risk_dimension_tables}
\title{Collate Quality Risks in tables for each division (adding list element names)}
\usage{
get_risk_dimension_tables(all_risk_list, new_list_names)
}
\arguments{
\item{all_risk_list}{List of quality risks (by division) containing sub lists of risks}

\item{new_list_names}{New names as values}
}
\value{
List of tables (named)
}
\description{
Function to produce table of risks for each division, with row totals for number of risks affecting each dimension and named list elements
}
