% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_risks.R
\name{get_sep_risks_all_div}
\alias{get_sep_risks_all_div}
\title{Extract Quality Risks for all divisions (with named list!)}
\usage{
get_sep_risks_all_div(file_path, division_names)
}
\arguments{
\item{file_path}{List of file paths}

\item{division_names}{Names as values}
}
\value{
List of tables
}
\description{
Function to extract tab names (for each division), extract Quality Risk tables for all divisions and name list elements
}
