% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_risks.R
\name{merge_all_div_risks}
\alias{merge_all_div_risks}
\title{Collate risks in tables for each division}
\usage{
merge_all_div_risks(all_div_sep_risk_list_named)
}
\arguments{
\item{all_div_sep_risk_list_named}{List of divisional risks containing sub lists}
}
\value{
List of tables (not named)
}
\description{
Function to produce table of risks for each division, with totals for risk to each dimension
}
