% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_risks.R
\name{merge_all_risks}
\alias{merge_all_risks}
\title{Collate Quality Risks in tables for each division}
\usage{
merge_all_risks(all_risk_list)
}
\arguments{
\item{all_risk_list}{List of quality risks (by division) containing sub lists of risks}
}
\value{
List of tables (not named)
}
\description{
Function to produce table of risks for each division, with row totals for number of risks affecting each dimension
}
