% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_metrics.R
\name{merge_metric_table}
\alias{merge_metric_table}
\title{Merge metrics for one division into one table}
\usage{
merge_metric_table(
  metric_list,
  config_file = "D:/Repos/quality_improvement_plans/config.yaml"
)
}
\arguments{
\item{metric_list}{List containing 4 metric tables for one division}

\item{config_file}{.yaml file containing config, with default settings (data directory, null values, metric ranges etc)}
}
\value{
Table with all metrics for one division
}
\description{
Function to merge all 4 metrics into one table for one division
}
