% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_risks.R
\name{merge_single_div_risks}
\alias{merge_single_div_risks}
\title{Collate Quality Risks for single division}
\usage{
merge_single_div_risks(all_div_sep_risk_list_named, division_order)
}
\arguments{
\item{all_div_sep_risk_list_named}{List of Quality divisions for single division}

\item{division_order}{Numeric vector from 1 to number of divisions}
}
\value{
Table with total column
}
\description{
Function to merge all Quality Risks for each division into one table, with row sums for number of risks affecting each dimension
}
