% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{read_config}
\alias{read_config}
\title{Reading in config file}
\usage{
read_config(config_file = "D:/Repos/quality_improvement_plans/config.yaml")
}
\arguments{
\item{config_file}{.yaml file containing config, with default settings (data directory, null values, metric ranges etc)}
}
\value{
List of variables
}
\description{
Function to read in configs of the QIPs, including directory and null_values
}
