% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_metrics.R
\name{recent_metric_table}
\alias{recent_metric_table}
\title{Get table with latest 2 months of data}
\usage{
recent_metric_table(
  all_metric_list,
  config_file = "D:/Repos/quality_improvement_plans/config.yaml"
)
}
\arguments{
\item{all_metric_list}{List by division containing 4 metric tables for each division}

\item{config_file}{.yaml file containing config, with default settings (data directory, null values, metric ranges etc)}
}
\value{
Single table with latest 2 months of data
}
\description{
Function to condense metric data from all divisions to figures only for latest 2 months
}
