% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_risks.R
\name{sum_full_cells}
\alias{sum_full_cells}
\title{Sum over non-empty rows}
\usage{
sum_full_cells(table)
}
\arguments{
\item{table}{Table with characters}
}
\value{
Table with added frequency column
}
\description{
Function to sum non-empty cells in rows (excluding first row as row name)
}
