% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_risks.R
\name{sum_no_risks}
\alias{sum_no_risks}
\title{Sum over numerical rows}
\usage{
sum_no_risks(table)
}
\arguments{
\item{table}{Table with character and numeric entries}
}
\value{
Table with added total column
}
\description{
Function to sum numerical cells in rows (excluding first row as row name)
}
