% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_risks.R
\name{sum_risks}
\alias{sum_risks}
\title{Sum over non-empty rows}
\usage{
sum_risks(table)
}
\arguments{
\item{table}{Table with character entries}
}
\value{
Table with added frequency column
}
\description{
Function to sum non-empty cells in rows (excluding first row as row name)
}
