% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_risks.R
\name{get_final_risk_dimension_table}
\alias{get_final_risk_dimension_table}
\title{Table of quality risks by division and dimension}
\usage{
get_final_risk_dimension_table(merged_risks)
}
\arguments{
\item{merged_risks}{List of risk tables by division}
}
\value{
Table
}
\description{
Function to produce table with all risks by dimension and division, with final % of Quality risks affecting each division
}
